Code refactor for signing. DRY
